+++ docs/sidebars.js
@@ -48,7 +48,8 @@ module.exports = {
       label: 'Developer Docs',
       items: [
         'dev-overview',
-        'dev-add-component'
+        'dev-add-component',
+        'dev-plugins'
       ],
     },
     {
@@ -69,6 +70,9 @@ module.exports = {
         'tutorials',
         'tutorial-visium',
         'tutorial-gh-pages',
+        'tutorial-plugin-view-type',
+        'tutorial-plugin-coordination-type',
+        'tutorial-plugin-file-type',
       ],
     },
   ],