+++ src/app/component-registry.js
@@ -1,4 +1,5 @@
 import { Component } from './constants';
+import { getPluginViewType, getPluginViewTypes } from './plugins';
 
 import DescriptionSubscriber from '../components/description/DescriptionSubscriber';
 import StatusSubscriber from '../components/status/StatusSubscriber';
@@ -13,7 +14,6 @@ import CellSetSizesPlotSubscriber from '../components/sets/CellSetSizesPlotSubsc
 import GenomicProfilesSubscriber from '../components/higlass/GenomicProfilesSubscriber';
 import ExpressionHistogramSubscriber from '../components/genes/ExpressionHistogramSubscriber';
 import CellSetExpressionPlotSubscriber from '../components/sets/CellSetExpressionPlotSubscriber';
-
 
 const registry = {
   [Component.DESCRIPTION]: DescriptionSubscriber,
@@ -34,9 +34,19 @@ const registry = {
 };
 
 export function getComponent(name) {
-  const component = registry[name];
+  let component = registry[name];
   if (component === undefined) {
-    throw new Error(`Could not find definition for "${name}" in registry.`);
+    component = getPluginViewType(name);
+    if (component === undefined) {
+      throw new Error(`Could not find definition for "${name}" in the core registry nor the plugin registry.`);
+    }
   }
-  return registry[name];
+  return component;
+}
+
+export function getViewTypes() {
+  return [
+    ...Object.keys(registry),
+    ...getPluginViewTypes(),
+  ];
 }